+++ src/TestSuite/ConsoleIntegrationTestCase.php
@@ -9,22 +9,20 @@
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/cakephp-thumber
  * @license     https://opensource.org/licenses/mit-license.php MIT License
- * @since       1.1.1
+ * @since       1.6.1
  */
 namespace Thumber\TestSuite;
 
 use Cake\Http\BaseApplication;
-use Cake\TestSuite\IntegrationTestCase as CakeIntegrationTestCase;
+use Cake\TestSuite\ConsoleIntegrationTestCase as CakeConsoleIntegrationTestCase;
 use Thumber\TestSuite\Traits\TestCaseTrait;
-use Thumber\ThumbsPathTrait;
 
 /**
- * IntegrationTestCase class
+ * ConsoleIntegrationTestCase class
  */
-abstract class IntegrationTestCase extends CakeIntegrationTestCase
+abstract class ConsoleIntegrationTestCase extends CakeConsoleIntegrationTestCase
 {
     use TestCaseTrait;
-    use ThumbsPathTrait;
 
     /**
      * Called before every test method
@@ -47,20 +45,5 @@ abstract class IntegrationTestCase extends CakeIntegrationTestCase
         $this->deleteAll();
 
         parent::tearDown();
-    }
-
-    /**
-     * Asserts content type
-     * @param string $type The content-type to check for
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     */
-    public function assertContentType($type, $message = '')
-    {
-        $this->skipIf(!version_compare(PHP_VERSION, '7.0', '>') &&
-            in_array($type, ['image/x-ms-bmp', 'image/vnd.adobe.photoshop']));
-
-        parent::assertContentType($type, $message);
     }
 }